Fixed bug local block parsing data size, check for local size incomplete
In the ZIP Format Specification (version 6.3.10) it is defined that for a local file header (section 4.3.7) flag bit 3 (section 4.4.4) may be set to indicate that the data size can be found in the data descriptor. A supporting implementation must in that case set crc32, compressed and uncompressed size to zero.

We can therefore assume that if C/U size or crc32 is not set to zero in the local file header, the given size is valid and can be used to determine the content size

I observed that some zip compression utils in fact include the size in the local file header while still setting bit 3.