deps: Use native lzma-rust2 instead of liblzma
This is a native port of the XZ's liblzma. The main advantage is, that the crate is native Rust and is also used by the 7z crate.

Performance should be equal to the liblzma crate, since I spend quite a bit of time to improve the performance.

I only used the single threaded version of the reader/writer, since parallelization was not enabled for liblzma. lzma-rust2 has multithreaded reader/writer though.

I had to remove the old bug report fix, since lzma_rust2 doesn't have the reported behavior. The test case also was obviously AI generated and useless.