Use native implementation of LZMA and XZ (#405)
* deps: Use native lzma-rust2 instead of liblzma

This is a native port of the XZ's liblzma. The main advantage is, that the crate is native Rust and is also used by the 7z crate.

Performance should be equal to the liblzma crate, since I spend quite a bit of time to improve the performance.

I only used the single threaded version of the reader/writer, since parallelization was not enabled for liblzma. lzma-rust2 has multithreaded reader/writer though.

I had to remove the old bug report fix, since lzma_rust2 doesn't have the reported behavior. The test case also was obviously AI generated and useless.

* fix: Properly implement LZMA decoding

The old implementation actually didn't worked. The test file uses STORE as compression. This makes sure that LZMA is properly implemented.

I remembered that I had to defer the initialization of reading the properties when implementing PPMd, so I did this here too.

* fix: Fix linter issue

* fix: Formatting of README.md

* fix: Reduce IO reads

* fix: Move XZ decoder and encoder to the heap

* fix: Fix code review issues

* fix: Fix UnwindSafe issue by using latest lzma-rust2 version

* Rephrase a comment in src/compression.rs

Signed-off-by: Chris Hennick <4961925+Pr0methean@users.noreply.github.com>

---------

Signed-off-by: Chris Hennick <4961925+Pr0methean@users.noreply.github.com>
Co-authored-by: Chris Hennick <4961925+Pr0methean@users.noreply.github.com>